Progress on lora masking
