+++ animatediff/nodes_conditioning.py
@@ -1,0 +1,67 @@
+import folder_paths
+
+from comfy.model_patcher import ModelPatcher
+import comfy.utils
+
+from .model_injection import ModelPatcherAndInjector
+
+# based on ComfyUI's nodes.py LoraLoader
+class MaskableLoraLoader:
+    def __init__(self):
+        self.loaded_lora = None
+
+    @classmethod
+    def INPUT_TYPES(s):
+        return {
+            "required": {
+                "model": ("MODEL",),
+                "clip": ("CLIP",),
+                "lora_name": (folder_paths.get_filename_list("loras"), ),
+                "strength_model": ("FLOAT", {"default": 1.0, "min": -20.0, "max": 20.0, "step": 0.01}),
+                "strength_clip": ("FLOAT", {"default": 1.0, "min": -20.0, "max": 20.0, "step": 0.01}),
+            }
+        }
+    
+    RETURN_TYPES = ("MODEL", "CLIP", "LORA_IDS")
+    CATEGORY = "Animate Diff 🎭🅐🅓/conditioning"
+    FUNCTION = "load_lora"
+
+    def load_lora(self, model: ModelPatcher, clip: ModelPatcher, lora_name: str, strength_model: float, strength_clip: float):
+        if strength_model == 0 and strength_clip == 0:
+            return (model, clip)
+        
+        lora_path = folder_paths.get_full_path("loras", lora_name)
+        lora = None
+        if self.loaded_lora is not None:
+            if self.loaded_lora[0] == lora_path:
+                lora = self.loaded_lora[1]
+            else:
+                temp = self.loaded_lora
+                self.loaded_lora = None
+                del temp
+        
+        if lora is None:
+            lora = comfy.utils.load_torch_file(lora_path, safe_load=True)
+            self.loaded_lora = (lora_path, lora)
+
+        model_lora, clip_lora = model.clone(), clip.clone()
+        return (model_lora, clip_lora)
+
+
+class MaskableLoraLoaderModelOnly(MaskableLoraLoader):
+    @classmethod
+    def INPUT_TYPES(s):
+        return {
+            "required": {
+                "model": ("MODEL",),
+                "lora_name": (folder_paths.get_filename_list("loras"), ),
+                "strength_model": ("FLOAT", {"default": 1.0, "min": -20.0, "max": 20.0, "step": 0.01}),
+            }
+        }
+
+    RETURN_TYPES = ("MODEL", "LORA_IDS")
+    CATEGORY = "Animate Diff 🎭🅐🅓/conditioning"
+    FUNCTION = "load_lora_model_only"
+
+    def load_lora_model_only(self, model: ModelPatcher, lora_name: str, strength_model: float):
+        return (self.load_lora(model, None, lora_name, strength_model, 0)[0],)
